% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseMAdata.R
\name{reverseMAdata}
\alias{reverseMAdata}
\title{reverseMAdata}
\usage{
reverseMAdata(X, M, Y, SEED = 1, nSimImai = 1000)
}
\arguments{
\item{X}{is the exposure}

\item{M}{is the mediator}

\item{Y}{is the outcome}

\item{SEED}{is the seed}

\item{nSimImai}{is the number of simulations for the mediate method}
}
\value{
a list of Estimates and p-values for the direct and indirect effects in situations with and without reverse causality.
}
\description{
Apply mediate function from the mediation package in scenarios of reverse causality.
}
\author{
Annie Thwing, Sharon Lutz
}
