% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseMRdata.R
\name{reverseMRdata}
\alias{reverseMRdata}
\title{reverseMRdata}
\usage{
reverseMRdata(nSNP = 3, X, M, Y, SEED = 1)
}
\arguments{
\item{nSNP}{the number of SNPS}

\item{X}{is a vector of SNP data, where each SNP is a column and the number of rows is the number of subjects}

\item{M}{is the mediator}

\item{Y}{is the outcome}

\item{SEED}{is the seed}
}
\value{
a list of Estimates and p-values for the direct and indirect effects in situations with and without reverse causality.
}
\description{
Apply MR methods from the MendelianRandomization package in scenarios of reverse causality.
}
\author{
Annie Thwing, Sharon Lutz
}
