% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseMRsim.R
\name{reverseMRsim}
\alias{reverseMRsim}
\title{reverseMRsim}
\usage{
reverseMRsim(n = 1000, nSNP = 3, MAF = c(0.2, 0.2, 0.2),
  gamma0 = 0, gammaX = c(0.2, 0.2, 0.2), varM = 1, beta0 = 0,
  betaM = c(0, 0.1), varY = 1, nSim = 100, plot.pdf = T,
  plot.name = "reverseMRsim.pdf", alpha_level = 0.05, SEED = 1)
}
\arguments{
\item{n}{is the sample size}

\item{nSNP}{is the number of SNPS to generate}

\item{MAF}{is a vector of the MAF of each SNP}

\item{gamma0}{is the intercept for M}

\item{gammaX}{is a vector of the associations of each SNP with M}

\item{varM}{is the variance of M}

\item{beta0}{is the intercept for Y}

\item{betaM}{is a vector of different associations of M with Y}

\item{varY}{is the variance of Y}

\item{nSim}{is the number of simulations to run}

\item{plot.pdf}{is T to output a plot, is F to not output a plot}

\item{plot.name}{is the name of the plot}

\item{alpha_level}{is the significance level}

\item{SEED}{is the seed}
}
\value{
a matrix of the power of three Mendelian Randomization approaches from the MendelianRandomization package to detect an effect of the mediator M on the outcome Y when M and Y are correctly specified and also when they are incorrectly specified (the true mediator is Y and the true outcome is M)
}
\description{
A function to simulate the performance of different MR methods from the MendelianRandomization package in scenarios of reverse causality.
}
\author{
Annie Thwing, Sharon Lutz
}
