% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_process_mediate.R
\name{simulate_and_mediate}
\alias{simulate_and_mediate}
\title{Simulate Data and Mediate Linear Models}
\usage{
simulate_and_mediate(med_model_vars)
}
\arguments{
\item{med_model_vars}{an instance of \code{MediateModelVariables}}
}
\value{
a \code{MediationProbValues} instance with values obtained by running mediate on forward and reverse linear models.
}
\description{
runs forward and reverse mediation on argument contents
}
