% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_process_mediate.R
\name{mediate_parallel}
\alias{mediate_parallel}
\title{Perform Mediation Using Parallel Processing}
\usage{
mediate_parallel(list_of_job_args, num_jobs = 1)
}
\arguments{
\item{list_of_job_args}{A list or matrix of list objects with named positions that can be used as inputs to simulate_and_mediatie}

\item{num_jobs}{number of parallel processes to use}
}
\value{
list of \code{MediationProbValues} objects
}
\description{
starts up multiple processes, configures their environment, and runs simulation and mediation analysis for each item in the input
}
