% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_and_mediate.R
\name{simulate_and_mediate}
\alias{simulate_and_mediate}
\title{Simulate Data and Mediate Linear Models}
\usage{
simulate_and_mediate(data_element)
}
\arguments{
\item{data_element}{list with named elements that will be used to assemble linear models to use with mediation}
}
\value{
a \code{MediationProbValues} instance with values obtained by running mediate on forward and reverse linear models.
}
\description{
runs forward and reverse mediation on argument contents
}
